Add support for @class, @extends and #dependsFile
This commits adds support for the following annotations in comments:

@class <CLASSNAME>
When a file contains this annotation <CLASSNAME> the file will be
considered to contain the definitation of the class with that name.

@extends <CLASSNAME>
When a file contains this annotation the file will be considered to
depend on the file that contains the definitation for the class with
that name.

\#dependsFile <FILENAME>
When a file contains this annotation the file will be considered to
depend on the file with the given name.